move to logging utils
